Completed assignment

+++ Binomial.java
@@ -4,12 +4,11 @@ public class Binomial {
 		//// Uncomment the version of binomial that you want to test
 
 		// Testing the basic binomial implementation:
-		//System.out.println(binomial1(Integer.parseInt(args[0]),
-		//Integer.parseInt(args[1])));
+		// System.out.println(binomial1(Integer.parseInt(args[0]),
+		// Integer.parseInt(args[1])));
 
 		// Testing the optimized binomial implementation:
-		System.out.println(binomial(Integer.parseInt(args[0]),
-		Integer.parseInt(args[1])));
+		System.out.println(binomial(Integer.parseInt(args[0]), Integer.parseInt(args[1])));
 	}
 
 	// Computes the Binomial function, basic version.
@@ -22,12 +21,12 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, / and then initializes all its
 		/// elements to -1. / It then calls binomial(n, k, memo), which does all the
 		/// heavy lifiting.
 		// Replace the following statement with your code.
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
 				memo[i][j] = -1;
@@ -36,7 +35,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
